// src/types/video-tutorial.ts
export interface VideoTutorial {
  id: number;
  register: string;
  guardianTutorRequest: string;
  tuitionBoard: string;
  tutorTuitionBoard: string;
  tutorPayment: string;
  tutorProfile: string;
  link1: string;
  link2: string;
  link3: string;
  others: string;
  createdAt: string;
  updatedAt: string;
}